requirements.txt
############################################################
# Core Scientific Stack
############################################################
numpy==1.26.4
pandas==2.1.4
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.2
pyarrow==15.0.0
############################################################
# Machine Learning
############################################################
scikit-learn==1.3.2
xgboost==2.1.1
imbalanced-learn==0.11.0
optuna==3.3.0
joblib==1.3.2
mlflow>=2.12

############################################################
# FastAPI / Web API
############################################################
fastapi
uvicorn
jinja2
requests==2.31.0
httpx==0.25.0
starlette==0.37.2
python-multipart==0.0.9

############################################################
# Data Contracts / Schema Validation
############################################################
pandera==0.19.2
pydantic==2.6.4

############################################################
# Data Versioning (DVC)
############################################################
dvc==3.46.0
dvc-gdrive==3.0.0          # If using Google Drive
dvc-s3==3.0.0              # If using S3 / MinIO
dvc-azure==3.0.0           # If using Azure Blob Storage

############################################################
# Config / Utilities
############################################################
pyyaml==6.0.1
tqdm==4.66.2
rich==13.7.0

############################################################
# Testing & Linting
############################################################
pytest==7.4.4
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
mypy==1.9.0

############################################################
# Development / Notebooks
############################################################
jupyterlab==4.1.0
ipykernel==6.28.0
evidently>=0.4.20
